split backend adapters and expose selector globally
