+++ karabo/imaging/backends/__init__.py
@@ -1,0 +1,11 @@
+"""Backend-specific imager adapters."""
+
+from .rascil_backend import RascilBackendConfig, RascilBackendImager
+from .sdp_backend import SdpImager, SdpImagerConfig
+
+__all__ = [
+    "RascilBackendImager",
+    "RascilBackendConfig",
+    "SdpImager",
+    "SdpImagerConfig",
+]
